feat(phase-11q1): implement Split CitySimulation into Sub-Systems and Introduce Sub-Interfaces
