+++ architecture/game-design/game-over-flow.md
@@ -7,10 +7,10 @@
   - **Month 3**: Game-over modal fires if deficit is still ≥ −50%.
   - **Streak-break feedback**: When the deficit recovers above −50% during an active streak (streak is broken), display a Normal toast: "Deficit streak broken — finances stabilizing." This confirms to the player that their intervention worked. The red flashing indicator is removed and simulation speed is not artificially constrained.
   - **Streak-reset rule**: The streak counter resets when any budget tick fires with deficit > −50% (even −49% breaks the streak). Streak month 1 re-fires a new CRITICAL toast only when the deficit falls to ≥ −50% again for the first time after a reset — do NOT re-fire the month-1 toast if the deficit has been continuously ≥ −50%.
-  - **Speed-reduction implementation**: The auto-slow to 1× on month 1 calls `CitySimulation::setSpeed(SpeedMultiplier::x1)` and fires an `AITOWN_HEADLESS`-compatible notification event. The player retains full speed selector control — this is a suggestion enforced once, not a persistent lock. If the player increases to 3× or 10×, they have acknowledged the risk.
-  - **Auto-slow implementation rule**: `CitySimulation::tick()` calls `setSpeed(SpeedMultiplier::x1)` unconditionally when it detects that the month-1 deficit streak threshold has been crossed — it does NOT check `GameMode` before doing so. Stopping simulation time is a pure simulation concern. `UIManager` is the caller that inspects `GameMode` before triggering UI-layer responses (e.g., toasts, the flashing budget indicator, and ultimately `transitionToGameOver()`). This separation is intentional: the simulation layer owns time; the UI layer owns presentation and mode-conditional gating.
+  - **Speed-reduction implementation**: The auto-slow to 1× on month 1 calls `SimTiming::setSpeed(SpeedMultiplier::x1)` (via `Population::doGameOverTick()`) and fires an `AITOWN_HEADLESS`-compatible notification event. The player retains full speed selector control — this is a suggestion enforced once, not a persistent lock. If the player increases to 3× or 10×, they have acknowledged the risk.
+  - **Auto-slow implementation rule**: `Population::doGameOverTick()` (called from `CitySimulation::doBudgetTick()`) calls `SimTiming::setSpeed(SpeedMultiplier::x1)` unconditionally when it detects that the month-1 deficit streak threshold has been crossed — it does NOT check `GameMode` before doing so. Stopping simulation time is a pure simulation concern. `UIManager` is the caller that inspects `GameMode` before triggering UI-layer responses (e.g., toasts, the flashing budget indicator, and ultimately `transitionToGameOver()`). This separation is intentional: the simulation layer owns time; the UI layer owns presentation and mode-conditional gating.
 
-After `CitySimulation::setSpeed(SpeedMultiplier::x1)` fires, the HUD speed selector MUST immediately reflect `x1` as the active speed. `UIManager` polls `ICitySimulation::getSpeedMultiplier()` via the speed-selector element each `UIManager::update()` call — the selector display updates automatically on the next frame without requiring a separate notification callback. `CitySimulation` does NOT fire a dedicated speed-change event — the `ICitySimulation::getSpeedMultiplier()` accessor is the sole communication channel between simulation and HUD for speed state. Phase 8 implementers must ensure the speed selector polls `getSpeedMultiplier()` every `UIManager::update()` call — not only on player input — so auto-slow events are reflected without UI-layer polling gaps.
+After `SimTiming::setSpeed(SpeedMultiplier::x1)` fires (via `Population::doGameOverTick()`), the HUD speed selector MUST immediately reflect `x1` as the active speed. `UIManager` polls `ICitySimulation::getSpeedMultiplier()` via the speed-selector element each `UIManager::update()` call — the selector display updates automatically on the next frame without requiring a separate notification callback. `CitySimulation` does NOT fire a dedicated speed-change event — the `ICitySimulation::getSpeedMultiplier()` accessor is the sole communication channel between simulation and HUD for speed state. Phase 8 implementers must ensure the speed selector polls `getSpeedMultiplier()` every `UIManager::update()` call — not only on player input — so auto-slow events are reflected without UI-layer polling gaps.
 
 - **ICitySimulation interface — deficit streak accessor**: The ICitySimulation interface must expose `getConsecutiveDeficitMonths() const -> int` so that the notification system can determine which progressive warning toast to fire (0 = no warning, 1 = "2 months to bankruptcy", 2 = "1 month to bankruptcy", 3+ = game-over trigger) without direct access to the concrete simulation state. This method returns 0 during the grace period. The method returns the count of completed deficit months at the END of the budget tick in which they were evaluated — i.e., after the budget tick that triggers the Nth deficit month fires, `getConsecutiveDeficitMonths()` returns N. At game-over trigger (month 3), the method returns 3. The grace-period return value of 0 is not a special override: it reflects the fact that deficit consequence evaluation is suppressed during the grace period, so the counter is never incremented.
 - **Sandbox mode**: Game-over does **not** apply in Sandbox mode (difficulty = any tier); Sandbox is open-ended
