+++ architecture/game-design/save-system.md
@@ -10,7 +10,7 @@
 - **Save in Scenario mode**: Scenario mode saves include scenario state (win condition progress, starting conditions). Scenario saves are stored in a dedicated `scenarios/` subdirectory.
 - **Population milestone serialization**: The save JSON MUST include a `population_milestone_fired` boolean array of exactly 5 elements, one per threshold (1K/10K/50K/100K/500K). Each element is `true` once the corresponding milestone has fired; `false` otherwise. On load, `CitySimulation` restores these flags to prevent already-fired milestones from re-triggering. Default-constructed value (new game) is `[false, false, false, false, false]`. See `SimulationConstants::population_milestone_threshold_1..5` for threshold values. Cross-reference: `implementation/phase-6.md` (Phase 11 promises — milestone flag serialization note).
 - **Speed multiplier serialization**: The save JSON MUST include a `speed_multiplier` field storing the integer representation of the `SpeedMultiplier` enum value active at save time. On load, `CitySimulation` restores this field so that the simulation resumes at the saved speed. After loading, `getSpeedMultiplier()` MUST return the value that was serialized. Default-constructed value (new game) is `SpeedMultiplier::x3` (the default starting speed per `architecture/game-design/simulation-time.md` — there is no `NORMAL` enum value). This field is verified by the `SaveSystem_RoundTrip_PreservesFullCityState` test in Phase 11.
-- **Building variant counter serialization** (Phase 11 scope): The save JSON MUST include a `building_variant_counters` array of exactly 9 integers, representing `CitySimulation::m_buildingVariantCounters` indexed by `zone * 3 + tier` (where `zone` = 0/1/2 for Res/Com/Ind and `tier` = 0/1/2 for Low/Med/High). On load, `CitySimulation` restores these counters so that round-robin variant selection resumes the existing sequence rather than restarting from 0. Without this, the first tile placed after a load always receives variant `01`, producing visible repetition when the pre-save sequence had already advanced to `02`. Default-constructed value (new game) is `[0, 0, 0, 0, 0, 0, 0, 0, 0]`. This field is verified by the `SaveSystem_RoundTrip_PreservesFullCityState` test in Phase 11. Cross-reference: `architecture/asset-standards/3d-model-standards.md` §91 ("Counter serialisation is Phase 11 scope").
+- **Building variant counter serialization** (Phase 11 scope): The save JSON MUST include a `building_variant_counters` array of exactly 9 integers, representing `Zoning::m_buildingVariantCounters` indexed by `zone * 3 + tier` (where `zone` = 0/1/2 for Res/Com/Ind and `tier` = 0/1/2 for Low/Med/High). On load, `CitySimulation` restores these counters so that round-robin variant selection resumes the existing sequence rather than restarting from 0. Without this, the first tile placed after a load always receives variant `01`, producing visible repetition when the pre-save sequence had already advanced to `02`. Default-constructed value (new game) is `[0, 0, 0, 0, 0, 0, 0, 0, 0]`. This field is verified by the `SaveSystem_RoundTrip_PreservesFullCityState` test in Phase 11. Cross-reference: `architecture/asset-standards/3d-model-standards.md` §91 ("Counter serialisation is Phase 11 scope").
 - **Tile footprint-origin serialization**: Each tile object in the `tiles` JSON array MUST include `fp_origin_x` and `fp_origin_z` integer fields. For origin tiles and 1×1 (Low-density) buildings, both are `-1`. For non-origin tiles in 2×2 (Medium) or 3×3 (High) footprints, they hold the coordinates of the footprint origin tile. **Omitting these fields causes double-building on load**: `rebuildCityFromSim()` uses `footprintOriginX == -1` as the criterion to place a building mesh; if all sub-tiles of a multi-tile building default to `-1`, a full mesh is spawned on each sub-tile, producing `N²` overlapping buildings. These fields are serialized as `"fp_origin_x"` / `"fp_origin_z"` and deserialized into `TileData::footprintOriginX` / `TileData::footprintOriginZ`. Backward-compat note: save files created before this field was added will have all tiles default to `-1` on load, which re-introduces the double-building bug for Medium and High density buildings in those saves.
 - **`LoadResult` enum**: `ISaveSystem::loadMostRecentSave()` returns a `LoadResult` value indicating the outcome of a load attempt. Defined values: `LoadResult::Ok` (save loaded and deserialized successfully), `LoadResult::NoSaveFound` (no save files exist — load should not be attempted when the Load Game button is correctly gated by `setSaveAvailable()`), `LoadResult::Corrupted` (save files exist but all fail checksum or schema validation). `UIManager` acts on `LoadResult::Ok` only; all other results leave the game in the Main Menu state.
 - **`loadMostRecentSave()` API**: `ISaveSystem::loadMostRecentSave()` loads the most recent valid save across all slots (auto-save and manual), sorted by timestamp, and deserializes it into the active `CitySimulation`. Returns `LoadResult::Ok` on success. Called by `UIManager` when `consumeLoadGameRequest()` returns `true`. This is the same underlying load operation used by the "Load Last Save" game-over flow.
