+++ tests/ui/demolition_input_test.cpp
@@ -142,7 +142,7 @@ protected:
         EXPECT_CALL(sim_, getCityRating()).WillRepeatedly(Return(CityRatingTier::Village));
         EXPECT_CALL(sim_, getTotalPopulation()).WillRepeatedly(Return(0));
         EXPECT_CALL(sim_, getSimulationTime()).WillRepeatedly(Return(SimulationTime{1, 1}));
-        EXPECT_CALL(sim_, getDemandPressurePct(_)).WillRepeatedly(Return(0.5f));
+        EXPECT_CALL(sim_, getZoneDemandFactor(_)).WillRepeatedly(Return(0.5f));
         EXPECT_CALL(sim_, getOutstandingBondUses()).WillRepeatedly(Return(2));
         EXPECT_CALL(sim_, getConsecutiveDeficitMonths()).WillRepeatedly(Return(0));
         EXPECT_CALL(sim_, pollPendingNotification(_)).WillRepeatedly(Return(false));
