+++ tests/ui/hud_grace_period_test.cpp
@@ -63,7 +63,7 @@ TEST_F(HUDGracePeriodTest, HUD_SecondNewGame_GracePeriodLabelVisible)
     ON_CALL(sim_, getCityRating()).WillByDefault(Return(CityRatingTier::Village));
     ON_CALL(sim_, getTotalPopulation()).WillByDefault(Return(0));
     ON_CALL(sim_, getSimulationTime()).WillByDefault(Return(SimulationTime{1, 1}));
-    ON_CALL(sim_, getDemandPressurePct(_)).WillByDefault(Return(0.0f));
+    ON_CALL(sim_, getZoneDemandFactor(_)).WillByDefault(Return(0.0f));
     ON_CALL(sim_, hasUndoPendingAction()).WillByDefault(Return(false));
     ON_CALL(sim_, getUndoExpiryTimeSeconds()).WillByDefault(Return(0.0));
     ON_CALL(sim_, isPaused()).WillByDefault(Return(false));
