+++ tests/ui/ui_manager_quit_test.cpp
@@ -68,7 +68,7 @@ protected:
         ON_CALL(sim_, getCityRating()).WillByDefault(Return(CityRatingTier::Village));
         ON_CALL(sim_, getTotalPopulation()).WillByDefault(Return(0));
         ON_CALL(sim_, getSimulationTime()).WillByDefault(Return(SimulationTime{1, 1}));
-        ON_CALL(sim_, getDemandPressurePct(_)).WillByDefault(Return(0.0f));
+        ON_CALL(sim_, getZoneDemandFactor(_)).WillByDefault(Return(0.0f));
         ON_CALL(sim_, hasUndoPendingAction()).WillByDefault(Return(false));
         ON_CALL(sim_, getUndoExpiryTimeSeconds()).WillByDefault(Return(0.0));
         ON_CALL(sim_, consumeBudgetTicks()).WillByDefault(Return(0));
@@ -335,7 +335,7 @@ protected:
         ON_CALL(sim_, getCityRating()).WillByDefault(Return(CityRatingTier::Village));
         ON_CALL(sim_, getTotalPopulation()).WillByDefault(Return(0));
         ON_CALL(sim_, getSimulationTime()).WillByDefault(Return(SimulationTime{1, 1}));
-        ON_CALL(sim_, getDemandPressurePct(_)).WillByDefault(Return(0.0f));
+        ON_CALL(sim_, getZoneDemandFactor(_)).WillByDefault(Return(0.0f));
         ON_CALL(sim_, hasUndoPendingAction()).WillByDefault(Return(false));
         ON_CALL(sim_, getUndoExpiryTimeSeconds()).WillByDefault(Return(0.0));
         ON_CALL(sim_, consumeBudgetTicks()).WillByDefault(Return(0));
