+++ tests/ui/uimanager_zone_overlay_test.cpp
@@ -126,7 +126,7 @@ protected:
         ON_CALL(sim_, getCityRating()).WillByDefault(Return(CityRatingTier::Village));
         ON_CALL(sim_, getTotalPopulation()).WillByDefault(Return(0));
         ON_CALL(sim_, getSimulationTime()).WillByDefault(Return(SimulationTime{1, 1}));
-        ON_CALL(sim_, getDemandPressurePct(_)).WillByDefault(Return(0.0f));
+        ON_CALL(sim_, getZoneDemandFactor(_)).WillByDefault(Return(0.0f));
         ON_CALL(sim_, hasUndoPendingAction()).WillByDefault(Return(false));
         ON_CALL(sim_, getUndoExpiryTimeSeconds()).WillByDefault(Return(0.0));
         ON_CALL(sim_, queryTile(_, _)).WillByDefault(Return(QueryResult{}));
